requirements.txt
pyautogui
pywinauto
pillow
python-xlib
psutil
requests
python-dotenv